Fixing more warnings

+++ src/configuration.h
@@ -448,7 +448,7 @@ static bool load_config_from_file(struct configuration *config, const char* path
             if (strncmp(val->key->data, CONFIG_KEY_IGNORE, val->key->len) != 0) {
                 break;
             }
-            const int cnt = config->ignore_players_count;
+            const short cnt = config->ignore_players_count;
             memcpy((char*)config->ignore_players[cnt],val->value->data, val->value->len);
             _trace("config::loaded_ignored_player: %s", config->ignore_players[cnt]);
             config->ignore_players_count++;
